implemented the create employee functionality and refactored first and last to just be name

+++ client/src/actions/employee.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { GET_EMPLOYEES, EMPLOYEES_ERROR, DELETE_EMPLOYEE } from './types';
+import { GET_EMPLOYEES, EMPLOYEES_ERROR, DELETE_EMPLOYEE, ADD_EMPLOYEE } from './types';
 
 export const getEmployees = () => async dispatch => {
     try {
@@ -18,12 +18,36 @@ export const getEmployees = () => async dispatch => {
 }
 
 export const deleteEmployee = id => async dispatch => {
+    if(window.confirm('Are you sure? This action cannot be undone!')) {
+        try {
+            await axios.delete(`/api/employees/${id}`)
+
+            dispatch({
+                type: DELETE_EMPLOYEE,
+                payload: id
+            })
+        } catch (err) {
+            dispatch({
+                type: EMPLOYEES_ERROR,
+                payload: { msg: err.response.statusText, status: err.response.status }
+            })
+        }
+    }
+}
+
+export const addEmployee = formData => async dispatch => {
+    const config = {
+        headers: {
+            'Content-type': 'application/json'
+        }
+    }
+
     try {
-        const res = await axios.delete(`/api/employees/${id}`)
+        const res = await axios.post(`/api/employees/`, formData, config)
 
         dispatch({
-            type: DELETE_EMPLOYEE,
-            payload: id
+            type: ADD_EMPLOYEE,
+            payload: res.data
         })
     } catch (err) {
         dispatch({
